requirements.txt
tensorflow==2.5.0
Keras==2.4.3
Pillow==7.1.2
numpy==1.19.5
matplotlib==3.2.2
pandas==1.1.5
scikit-image==0.16.2
scikit-learn==0.22.2.post1
six==1.15.0
sklearn==0.0
sklearn-pandas==1.8.0